requirements.txt
openai==1.54.0
requests==2.31.0
edge-tts==6.1.12
moviepy==1.0.3
Pillow==10.4.0
python-dotenv==1.0.0
pydub==0.25.1
